+++ ruby/recheck-rails/changelog.md
@@ -1,5 +1,8 @@
 ## ## [Unreleased]
 
+* setup: Generate a Checker from ActiveRecord validations.
+  Only PresenceValidation to start.
+
 ## 0.4.0 - 2025-07-16
 
 * Public beta.